requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.25
chromadb==1.0.15
pypdf==5.2.0
python-dotenv==1.0.1
openai==1.97.1
sentence-transformers==3.2.1
torch==2.5.1
fastapi==0.115.6
uvicorn[standard]==0.35.0 